Compiles with -Wconversion

+++ src/Bpp/Seq/Io/Maf/AlignmentFilterMafIterator.cpp
@@ -86,7 +86,7 @@ MafBlock* AlignmentFilterMafIterator::analyseCurrentBlock_() throw (Exception)
 
 
       //First we create a mask:
-      vector<unsigned int> pos;
+      vector<size_t> pos;
       vector<int> col(nr);
       //Reset window:
       window_.clear();
@@ -136,7 +136,7 @@ MafBlock* AlignmentFilterMafIterator::analyseCurrentBlock_() throw (Exception)
         }
       
         //Move forward:
-        for (unsigned int k = 0; k < step_; ++k) {
+        for (size_t k = 0; k < step_; ++k) {
           for (size_t j = 0; j < nr; ++j) {
             col[j] = aln[j][i];
           }
@@ -204,7 +204,7 @@ MafBlock* AlignmentFilterMafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* newBlock = new MafBlock();
             newBlock->setScore(block->getScore());
             newBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* subseq;
               if (i == 0) {
                 subseq = block->getSequence(j).subSequence(0, pos[i]);
@@ -221,7 +221,7 @@ MafBlock* AlignmentFilterMafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* outBlock = new MafBlock();
             outBlock->setScore(block->getScore());
             outBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* outseq = block->getSequence(j).subSequence(pos[i], pos[i + 1] - pos[i]);
               outBlock->addSequence(*outseq);
               delete outseq;
@@ -234,7 +234,7 @@ MafBlock* AlignmentFilterMafIterator::analyseCurrentBlock_() throw (Exception)
           MafBlock* newBlock = new MafBlock();
           newBlock->setScore(block->getScore());
           newBlock->setPass(block->getPass());
-          for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+          for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
             MafSequence* subseq;
             subseq = block->getSequence(j).subSequence(pos[pos.size() - 1], block->getNumberOfSites() - pos[pos.size() - 1]);
             newBlock->addSequence(*subseq);
@@ -290,7 +290,7 @@ MafBlock* AlignmentFilter2MafIterator::analyseCurrentBlock_() throw (Exception)
         }
       }
       //First we create a mask:
-      vector<unsigned int> pos;
+      vector<size_t> pos;
       vector<bool> col(nr);
       //Reset window:
       window_.clear();
@@ -341,7 +341,7 @@ MafBlock* AlignmentFilter2MafIterator::analyseCurrentBlock_() throw (Exception)
         }
       
         //Move forward:
-        for (unsigned int k = 0; k < step_; ++k) {
+        for (size_t k = 0; k < step_; ++k) {
           for (size_t j = 0; j < nr; ++j) {
             col[j] = (aln[j][i] == gap || aln[j][i] == unk);
           }
@@ -412,7 +412,7 @@ MafBlock* AlignmentFilter2MafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* newBlock = new MafBlock();
             newBlock->setScore(block->getScore());
             newBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* subseq;
               if (i == 0) {
                 subseq = block->getSequence(j).subSequence(0, pos[i]);
@@ -429,7 +429,7 @@ MafBlock* AlignmentFilter2MafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* outBlock = new MafBlock();
             outBlock->setScore(block->getScore());
             outBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* outseq = block->getSequence(j).subSequence(pos[i], pos[i + 1] - pos[i]);
               outBlock->addSequence(*outseq);
               delete outseq;
@@ -442,7 +442,7 @@ MafBlock* AlignmentFilter2MafIterator::analyseCurrentBlock_() throw (Exception)
           MafBlock* newBlock = new MafBlock();
           newBlock->setScore(block->getScore());
           newBlock->setPass(block->getPass());
-          for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+          for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
             MafSequence* subseq;
             subseq = block->getSequence(j).subSequence(pos[pos.size() - 1], block->getNumberOfSites() - pos[pos.size() - 1]);
             newBlock->addSequence(*subseq);
