+++ src/Bpp/Seq/Io/Maf/MaskFilterMafIterator.cpp
@@ -69,7 +69,7 @@ MafBlock* MaskFilterMafIterator::analyseCurrentBlock_() throw (Exception)
       size_t nr = aln.size();
       size_t nc = block->getNumberOfSites();
       //First we create a mask:
-      vector<unsigned int> pos;
+      vector<size_t> pos;
       vector<bool> col(nr);
       //Reset window:
       window_.clear();
@@ -109,7 +109,7 @@ MafBlock* MaskFilterMafIterator::analyseCurrentBlock_() throw (Exception)
         }
       
         //Move forward:
-        for (unsigned int k = 0; k < step_; ++k) {
+        for (size_t k = 0; k < step_; ++k) {
           for (size_t j = 0; j < nr; ++j) {
             col[j] = aln[j][i];
           }  
@@ -169,7 +169,7 @@ MafBlock* MaskFilterMafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* newBlock = new MafBlock();
             newBlock->setScore(block->getScore());
             newBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* subseq;
               if (i == 0) {
                 subseq = block->getSequence(j).subSequence(0, pos[i]);
@@ -186,7 +186,7 @@ MafBlock* MaskFilterMafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* outBlock = new MafBlock();
             outBlock->setScore(block->getScore());
             outBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* outseq = block->getSequence(j).subSequence(pos[i], pos[i + 1] - pos[i]);
               outBlock->addSequence(*outseq);
               delete outseq;
@@ -199,7 +199,7 @@ MafBlock* MaskFilterMafIterator::analyseCurrentBlock_() throw (Exception)
           MafBlock* newBlock = new MafBlock();
           newBlock->setScore(block->getScore());
           newBlock->setPass(block->getPass());
-          for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+          for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
             MafSequence* subseq;
             subseq = block->getSequence(j).subSequence(pos[pos.size() - 1], block->getNumberOfSites() - pos[pos.size() - 1]);
             newBlock->addSequence(*subseq);
